Update upgrade version feature
- Rename the feature to upgrade
- Change to takes packages from standard input
- Put test fixtures as files in tests/fixtures
- Add argument with the versions to transform to/from
- Avoid except Exception and other code updates
