+++ ocdskit/cli/__main__.py
@@ -22,7 +22,7 @@ COMMAND_MODULES = (
     'ocdskit.cli.commands.split_record_packages',
     'ocdskit.cli.commands.split_release_packages',
     'ocdskit.cli.commands.tabulate',
-    'ocdskit.cli.commands.update_to_v1_1',
+    'ocdskit.cli.commands.upgrade',
     'ocdskit.cli.commands.validate',
 )
