+++ tests/fixtures/realdata/release_package_1_0_contracts.json
@@ -1,0 +1,1 @@
+{"uri":"https://contrataciones.gov.py/datos/id/contratos/246807-11-setiembre-srl-4","publishedDate":"2018-12-18T13:20:42Z","publisher":{"name":"DNCP - Paraguay","legalName":"Dirección Nacional de Contrataciones Públicas, Paraguay","uri":"https://contrataciones.gov.py/datos"},"license":"https://creativecommons.org/licenses/by/4.0/","publicationPolicy":"https://www.contrataciones.gov.py/datos/legal","releases":[{"language":"es","ocid":"ocds-03ad3f-246807","id":"246807-11-setiembre-srl-4-contract","date":"2018-12-18T13:20:42Z","tag":["contract"],"initiationType":"tender","contracts":[{"implementation":{"transactions":[{"id":"92885","amount":{"amount":2372509,"currency":"PYG"},"date":"2012-05-31T00:00:00Z","providerOrganization":{"id":"PY-PGN-12-8-1000000","legalName":"1000000-DIRECCION GENERAL DE ADMINISTRACION Y FINANZAS","scheme":"PY-PGN"},"receiverOrganization":{"id":"PY-RUC-80017437-2","legalName":"NUCLEO S.A.","scheme":"PY-PGN"}}]},"id":"246807-11-setiembre-srl-4","awardID":"246807-adqusicion-guantes-otros-insumos-medicos","dncpContractCode":"CD-13003-13-70684","title":"ADQUSICION DE GUANTES Y OTROS INSUMOS MEDICOS ","status":"active","value":{"amount":17875000,"currency":"PYG"},"items":[{"description":"Placa Radiografica Medica","id":"M99UFYWPOag%3D","classification":{"schema":"Catálogo de productos, bienes y servicios de la Dirección Nacional de Contrataciones","id":"42201810-001","description":"Catálogo de productos, bienes y servicios de la Dirección Nacional de Contrataciones","uri":"https://www.contrataciones.gov.py/datos/api/v2/doc/catalogo/nivel-5/42201810-001"},"quantity":10,"unit":{"name":"Unidad","value":{"amount":445000,"currency":"PYG"}}},{"description":"Placa Radiografica Medica","id":"7SmiEGR9Oes%3D","classification":{"schema":"Catálogo de productos, bienes y servicios de la Dirección Nacional de Contrataciones","id":"42201810-001","description":"Catálogo de productos, bienes y servicios de la Dirección Nacional de Contrataciones","uri":"https://www.contrataciones.gov.py/datos/api/v2/doc/catalogo/nivel-5/42201810-001"},"quantity":10,"unit":{"name":"Unidad","value":{"amount":850000,"currency":"PYG"}}},{"description":"Liquido fijador para placas radiograficas","id":"hPQ2FURf7RQ%3D","classification":{"schema":"Catálogo de productos, bienes y servicios de la Dirección Nacional de Contrataciones","id":"42203708-002","description":"Catálogo de productos, bienes y servicios de la Dirección Nacional de Contrataciones","uri":"https://www.contrataciones.gov.py/datos/api/v2/doc/catalogo/nivel-5/42203708-002"},"quantity":20,"unit":{"name":"Unidad","value":{"amount":110000,"currency":"PYG"}}},{"description":"Liquido revelador para placas radiograficas","id":"3LykMkkFXqU%3D","classification":{"schema":"Catálogo de productos, bienes y servicios de la Dirección Nacional de Contrataciones","id":"42203704-001","description":"Catálogo de productos, bienes y servicios de la Dirección Nacional de Contrataciones","uri":"https://www.contrataciones.gov.py/datos/api/v2/doc/catalogo/nivel-5/42203704-001"},"quantity":20,"unit":{"name":"Unidad","value":{"amount":110000,"currency":"PYG"}}},{"description":"Pinza para laboratorio","id":"2iT5c5pHD9g%3D","classification":{"schema":"Catálogo de productos, bienes y servicios de la Dirección Nacional de Contrataciones","id":"41122404-001","description":"Catálogo de productos, bienes y servicios de la Dirección Nacional de Contrataciones","uri":"https://www.contrataciones.gov.py/datos/api/v2/doc/catalogo/nivel-5/41122404-001"},"quantity":5,"unit":{"name":"Unidad","value":{"amount":105000,"currency":"PYG"}}}],"dateSigned":"2013-04-05T12:00:00Z","suppliers":[{"name":"11 DE SETIEMBRE SRL"}],"period":{"endDate":"2013-04-15T12:00:00Z","startDate":"2013-04-05T12:00:00Z"},"url":"https://www.contrataciones.gov.py/datos/id/contratos/246807-11-setiembre-srl-4"}]}]}
